Safari-609.1.11.1/LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin-redirect-expected.txt

Verifying that navigator.sendBeacon() to non-CORS cross-origin redirect fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("http://127.0.0.1:8080/navigation/resources/redirection-response.php?status=302&target=/non-existent.php", "CrossOrigin"); is true
PASS successfullyParsed is true

TEST COMPLETE
